+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+FEATURES:
+* Output map of `tfe_run_trigger` IDs which link to the source workspace
+
 ## [5.4.0] - 2025-01-12
 
 FEATURES: